+++ tests/test_helauth/test_profile_client.py
@@ -8,7 +8,7 @@ import pytest
 from django.conf import settings
 from graphene_django_extensions.testing import parametrize_helper
 
-from tilavarauspalvelu.utils.helauth.clients import HelsinkiProfileClient
+from tilavarauspalvelu.integrations.helauth.clients import HelsinkiProfileClient
 from utils.external_service.errors import (
     ExternalServiceError,
     ExternalServiceParseJSONError,
