+++ tests/test_utils/test_jwt_decoding.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from tilavarauspalvelu.utils.helauth.utils import get_jwt_payload
+from tilavarauspalvelu.integrations.helauth.utils import get_jwt_payload
 
 
 def test_get_jwt_payload():
