+++ tilavarauspalvelu/api/graphql/types/helsinki_profile/types.py
@@ -8,10 +8,10 @@ from graphql import GraphQLError
 from query_optimizer.selections import get_field_selections
 
 from tilavarauspalvelu.api.graphql.extensions import error_codes
+from tilavarauspalvelu.integrations.helauth.clients import HelsinkiProfileClient
+from tilavarauspalvelu.integrations.helauth.typing import LoginMethod, UserProfileInfo
 from tilavarauspalvelu.models import Application, Reservation
 from tilavarauspalvelu.tasks import save_personal_info_view_log
-from tilavarauspalvelu.utils.helauth.clients import HelsinkiProfileClient
-from tilavarauspalvelu.utils.helauth.typing import LoginMethod, UserProfileInfo
 
 if TYPE_CHECKING:
     import datetime
