+++ tilavarauspalvelu/integrations/helauth/clients.py
@@ -14,7 +14,7 @@ from social_django.models import DjangoStorage
 from social_django.strategy import DjangoStrategy
 
 from config.auth import ProxyTunnistamoOIDCAuthBackend
-from tilavarauspalvelu.utils.helauth.parsers import ProfileDataParser
+from tilavarauspalvelu.integrations.helauth.parsers import ProfileDataParser
 from utils.date_utils import DEFAULT_TIMEZONE, local_datetime
 from utils.external_service.base_external_service_client import BaseExternalServiceClient
 from utils.external_service.errors import ExternalServiceError, ExternalServiceRequestError
@@ -23,8 +23,7 @@ from utils.sentry import SentryLogger
 if TYPE_CHECKING:
     from requests import Response
 
-    from tilavarauspalvelu.typing import AnyUser, SessionMapping
-    from tilavarauspalvelu.utils.helauth.typing import (
+    from tilavarauspalvelu.integrations.helauth.typing import (
         BirthdayInfo,
         ExtraData,
         ProfileData,
@@ -33,6 +32,7 @@ if TYPE_CHECKING:
         TokenResponse,
         UserProfileInfo,
     )
+    from tilavarauspalvelu.typing import AnyUser, SessionMapping
 
 __all__ = [
     "HelsinkiProfileClient",
