+++ tilavarauspalvelu/integrations/helauth/typing.py
@@ -75,7 +75,7 @@ class IDToken:
 
     @classmethod
     def from_string(cls, token: str) -> IDToken | None:
-        from tilavarauspalvelu.utils.helauth.utils import get_jwt_payload
+        from tilavarauspalvelu.integrations.helauth.utils import get_jwt_payload
 
         try:
             payload = get_jwt_payload(token)
