+++ tilavarauspalvelu/models/user/model.py
@@ -12,8 +12,8 @@ from helsinki_gdpr.models import SerializableMixin
 from helusers.models import AbstractUser
 
 from tilavarauspalvelu.enums import ReservationNotification, UserRoleChoice
-from tilavarauspalvelu.utils.helauth.typing import IDToken
-from tilavarauspalvelu.utils.helauth.utils import get_jwt_payload
+from tilavarauspalvelu.integrations.helauth.typing import IDToken
+from tilavarauspalvelu.integrations.helauth.utils import get_jwt_payload
 from tilavarauspalvelu.utils.permission_resolver import PermissionResolver
 from utils.date_utils import DEFAULT_TIMEZONE
 
@@ -25,8 +25,8 @@ if TYPE_CHECKING:
     from social_django.models import UserSocialAuth
 
     from tilavarauspalvelu.enums import UserPermissionChoice
+    from tilavarauspalvelu.integrations.helauth.typing import ExtraData
     from tilavarauspalvelu.models import UnitRole
-    from tilavarauspalvelu.utils.helauth.typing import ExtraData
 
     from .actions import UserActions
